requirements-dev.txt
-r requirements.txt
pytest
pytest-cov
pytest-django
coverage
